chaos: move Model Shim API to core from ast
