+++ packages/@ts-graphviz/ast/src/index.ts
@@ -4,4 +4,3 @@
 export * from './types.js';
 export * from './builder/index.js';
 export * from './dot-shim/index.js';
-export * from './model-shim/index.js';
