+++ packages/@ts-graphviz/core/package.json
@@ -31,6 +31,7 @@
   },
   "homepage": "https://github.com/ts-graphviz/ts-graphviz#readme",
   "dependencies": {
+    "@ts-graphviz/ast": "^2.0.0",
     "@ts-graphviz/common": "^2.0.0"
   },
   "devDependencies": {
