+++ packages/@ts-graphviz/core/src/index.ts
@@ -1,1 +1,2 @@
 export * from './models/index.js';
+export * from './model-shim/index.js';
