+++ packages/@ts-graphviz/core/src/model-shim/from-model/plugins/AttributeListPlugin.ts
@@ -1,8 +1,8 @@
 import { AttributeListModel, getASTType } from '@ts-graphviz/common';
 import { ConvertFromModelPlugin } from '../types.js';
 import { convertAttribute } from './utils/index.js';
-import { createElement } from '../../../builder/create-element.js';
 import { getAttributeListKind } from '@ts-graphviz/common';
+import { createElement } from '@ts-graphviz/ast';
 
 export const AttributeListPlugin: ConvertFromModelPlugin<AttributeListModel> = {
   match(model) {
