+++ packages/@ts-graphviz/core/src/model-shim/from-model/plugins/index.ts
