+++ packages/@ts-graphviz/core/src/model-shim/index.ts
