+++ packages/@ts-graphviz/core/src/model-shim/to-model/converter.ts
