+++ packages/@ts-graphviz/core/src/model-shim/to-model/plugins/GraphPlugin.ts
@@ -1,4 +1,4 @@
-import { GraphASTNode } from '../../../types.js';
+import { GraphASTNode } from '@ts-graphviz/ast';
 import { ConvertToModelPlugin } from '../types.js';
 import { applyStatements } from './utils/apply-statments.js';
